requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.25
pandas==2.1.4
python-binance==1.0.19
websocket-client==1.7.0
requests==2.31.0
numpy==1.26.2